moved runners to their own files, changed os execution to realative to installed dir, updated installed to copy files
