--- BashRunner.sh
@@ -1,11 +1,0 @@
-#!/bin/bash
-
-echo BashRunner Ran!
-
-echo $'\n' >> testFile.txt
-
-echo $'Text Written From BashRunner!' >> testFile.txt
-
-echo "test file written, go rubyRunner!"
-
-ruby rubyRunner.rb
