+++ runners/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "radrouterun",
+  "version": "1.0.0",
+  "description": "",
+  "main": "noder.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/shyaboi/rad-route-runner.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/shyaboi/rad-route-runner/issues"
+  },
+  "homepage": "https://github.com/shyaboi/rad-route-runner#readme"
+}
